requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
PyPDF2
python-docx
beautifulsoup4
